app/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
sqlalchemy==1.4.50
asyncpg==0.28.0
psycopg2-binary==2.9.9
redis==5.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic[email]==2.4.2
email-validator==2.1.0
python-multipart==0.0.6
google-cloud-storage==2.12.0
